Add synth price and balance

+++ src/IMarket.sol
@@ -3,5 +3,5 @@ pragma solidity ^0.8.13;
 
 // Someone else's code here.
 interface IMarket {
-  function balance() external view returns (uint);
+    function balance() external view returns (int256);
 }
